GUI połączone z klientem
