+++ warcaby/src/main/java/tp/README.txt
@@ -1,0 +1,8 @@
+Dla odpalenia javafx może się przydać ustawienie poniższego kawałku kodu w ustawieniach "add VM options":
+
+    --module-path
+    "C:\Users\hnatiuk\Downloads\openjfx-19_windows-x64_bin-sdk\javafx-sdk-19\lib"
+    --add-modules
+    javafx.controls,javafx.fxml
+
+W module-path warto wpisać własną ścieżkę do folderu lib.
